requirements.txt
numpy==1.21.0
spacy==3.7.6
flask==2.0.1
werkzeug==2.0.1
pytest==6.2.5
python-docx==0.8.11
pdfplumber==0.7.4
typer==0.3.2
beautifulsoup4==4.12.2
requests==2.31.0
scikit-learn==1.2.2
SQLAlchemy==1.4.31
PyPDF2==1.26.0